Add i18n support with German translations
