+++ locales/de/steps/15.json
@@ -1,0 +1,14 @@
+{
+  "id": "15",
+  "title": "DE Feedback Loop",
+  "description": "DE The feedback loop closes the cycle by incorporating insights and user feedback back into the RAG pipeline, enabling continuous improvement over time.",
+  "whatItDoes": "DE This process takes learnings from monitoring, user feedback, and evaluation to make iterative improvements throughout the pipeline. It includes adding missing content, refining chunking strategies, updating embeddings, fine-tuning models, and enhancing prompts based on real-world usage patterns and identified gaps.",
+  "whyItMatters": "DE No RAG system is perfect from the start. The feedback loop is how your system evolves and adapts to changing needs and expectations. By actively incorporating feedback, you can address content gaps, improve retrieval accuracy, enhance answer quality, and ensure your system remains relevant and effective as requirements change.",
+  "challenges": [
+    "DE Collecting meaningful feedback from users who may not always provide explicit ratings",
+    "DE Analyzing diverse feedback data to identify actionable improvements",
+    "DE Avoiding regressions when implementing changes to fix specific issues",
+    "DE Balancing qualitative insights with quantitative metrics",
+    "DE Establishing organizational processes for continuous improvement"
+  ]
+}
