+++ locales/de/steps/6.json
@@ -1,0 +1,14 @@
+{
+  "id": "6",
+  "title": "DE Embedding Creation",
+  "description": "DE Embedding creation transforms text chunks into numerical vector representations that capture semantic meaning.",
+  "whatItDoes": "DE Embedding models convert text into high-dimensional vectors where similar texts are positioned closer together in the vector space. These embeddings enable semantic search and similarity matching.",
+  "whyItMatters": "DE The quality of your embeddings directly affects retrieval accuracy. Better embeddings capture more nuanced semantic relationships and lead to more relevant search results.",
+  "challenges": [
+    "DE Selecting appropriate embedding models for your domain",
+    "DE Handling multilingual content effectively",
+    "DE Managing embedding dimensionality and computational costs",
+    "DE Dealing with embedding model limitations and biases",
+    "DE Keeping embeddings up-to-date with model advancements"
+  ]
+}
