+++ locales/en/steps/12.json
@@ -1,0 +1,14 @@
+{
+  "id": "12",
+  "title": "Answer Generation",
+  "description": "Answer generation is where the LLM produces a response based on the user's query and the retrieved context, synthesizing information into a coherent answer.",
+  "whatItDoes": "The LLM takes the assembled context and user query to generate a natural language response. It synthesizes relevant information from the context, formats the answer appropriately, and delivers the final output to the user.",
+  "whyItMatters": "This is the culmination of the RAG pipeline where value is delivered to the user. A well-performed answer generation produces responses that are correct (grounded in facts), complete, and clear, significantly reducing hallucinations compared to using the LLM alone.",
+  "challenges": [
+    "Preventing hallucinations even with provided context",
+    "Ensuring the model uses the context rather than prior knowledge",
+    "Balancing verbosity and relevance in responses",
+    "Handling cases where the retrieved context doesn't contain the answer",
+    "Preventing information leakage from sensitive context"
+  ]
+}
