+++ locales/en/steps/4.json
@@ -1,0 +1,14 @@
+{
+  "id": "4",
+  "title": "Indexing",
+  "description": "Indexing organizes your document collection for efficient storage and retrieval, making it possible to quickly find relevant information.",
+  "whatItDoes": "Indexing creates optimized data structures that allow for fast searching and retrieval of documents. It maps terms, vectors, or other features to the documents that contain them.",
+  "whyItMatters": "Effective indexing is crucial for the performance and scalability of your RAG system. It enables quick retrieval of relevant documents from potentially massive collections.",
+  "challenges": [
+    "Choosing the right index type for your specific use case",
+    "Balancing index size with query performance",
+    "Handling index updates as documents change",
+    "Scaling indexes to large document collections",
+    "Optimizing for different query patterns"
+  ]
+}
