+++ locales/en/steps/5.json
@@ -1,0 +1,14 @@
+{
+  "id": "5",
+  "title": "Chunking Design",
+  "description": "Chunking breaks down documents into smaller, manageable pieces that can be individually embedded and retrieved.",
+  "whatItDoes": "Chunking splits documents into smaller segments based on various strategies like size, semantic boundaries, or structural elements. These chunks become the basic units for embedding and retrieval.",
+  "whyItMatters": "The way you chunk documents significantly impacts retrieval quality. Good chunking preserves context, reduces noise, and ensures that relevant information is retrieved together.",
+  "challenges": [
+    "Determining optimal chunk size for your specific use case",
+    "Preserving context across chunk boundaries",
+    "Handling documents with varying structures",
+    "Balancing chunk granularity with retrieval precision",
+    "Managing overlapping chunks to preserve context"
+  ]
+}
