+++ locales/en/steps/6.json
@@ -1,0 +1,14 @@
+{
+  "id": "6",
+  "title": "Embedding Creation",
+  "description": "Embedding creation transforms text chunks into numerical vector representations that capture semantic meaning.",
+  "whatItDoes": "Embedding models convert text into high-dimensional vectors where similar texts are positioned closer together in the vector space. These embeddings enable semantic search and similarity matching.",
+  "whyItMatters": "The quality of your embeddings directly affects retrieval accuracy. Better embeddings capture more nuanced semantic relationships and lead to more relevant search results.",
+  "challenges": [
+    "Selecting appropriate embedding models for your domain",
+    "Handling multilingual content effectively",
+    "Managing embedding dimensionality and computational costs",
+    "Dealing with embedding model limitations and biases",
+    "Keeping embeddings up-to-date with model advancements"
+  ]
+}
