+++ next.config.mjs
@@ -1,5 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  i18n: {
+    locales: ["en", "de"],
+    defaultLocale: "en",
+  },
   eslint: {
     ignoreDuringBuilds: true,
   },
